Refactor B-Tree implementation: enhance comments, improve code readability, and update CMake configuration for better clarity

+++ playground.cpp
@@ -1,8 +1,13 @@
+// playground.cpp
+// Programa de demonstração e testes básicos da implementação da B-Tree.
+// Este ficheiro contém uma implementação completa (standalone) da B-Tree
+// seguida por um main que insere alguns números aleatórios e imprime a árvore.
+
 #include <iostream>
-#include <print>
 #include <random>
 #include <vector>
 
+// Estrutura do nó (duplicada aqui para que o playground seja standalone)
 struct bTreeNode
 {
     std::vector<int> keys;
@@ -10,7 +15,7 @@ struct bTreeNode
     bool isLeaf;
     int t;
     int MaxKeys() const { return 2 * t - 1; }
-    int MiddleKey() const { return t - 1; }  
+    int MiddleKey() const { return t - 1; }
 
 public:
     bTreeNode(int t, bool isLeaf)
@@ -18,11 +23,13 @@ public:
         this->t = t;
         this->isLeaf = isLeaf;
 
+        // Reservamos capacidade para melhorar performance nas inserções
         keys.reserve(MaxKeys());
         children.reserve(2 * t);
     }
 };
 
+// Classe B-Tree (implementação mínima, igual à usada no código principal)
 class bTree
 {
     bTreeNode* root;
@@ -39,7 +46,6 @@ public:
     void printTree();
     void traverse();
 
-
 private:
     void traverseNode(bTreeNode* node);
     void insertNonFull(bTreeNode* node, int key);
@@ -49,40 +55,42 @@ private:
 
 // ---------- Implementações ----------
 
+// Insere uma chave na árvore; trata crescimento da altura quando a raiz está cheia
 void bTree::insert(int key)
 {
-    if(root->keys.size() == root->MaxKeys()){
+    if (root->keys.size() == root->MaxKeys()) {
         bTreeNode* newRoot = new bTreeNode(t, false);
         newRoot->children.push_back(root);
         splitChild(newRoot, 0);
         int i = 0;
-        if(newRoot->keys[0] < key){
+        if (newRoot->keys[0] < key) {
             i++;
         }
         insertNonFull(newRoot->children[i], key);
         root = newRoot;
-    }else{
+    } else {
         insertNonFull(root, key);
     }
 }
 
+// Insere em nó não cheio (comportamento comum usado pela B-Tree)
 void bTree::insertNonFull(bTreeNode* node, int key)
 {
-    if(node->isLeaf){
-        int i = node->keys.size() -1;
-        while(i >= 0 && key < node->keys[i]){
+    if (node->isLeaf) {
+        int i = (int)node->keys.size() - 1;
+        while (i >= 0 && key < node->keys[i]) {
             i--;
         }
-        node->keys.insert(node->keys.begin() + i +1, key);
-    }else{
-        int i = node->keys.size() -1;
-        while(i >=0 && key < node->keys[i]){
+        node->keys.insert(node->keys.begin() + i + 1, key);
+    } else {
+        int i = (int)node->keys.size() - 1;
+        while (i >= 0 && key < node->keys[i]) {
             i--;
         }
         i++;
-        if(node->children[i]->keys.size() == node->children[i]->MaxKeys()){
+        if (node->children[i]->keys.size() == node->children[i]->MaxKeys()) {
             splitChild(node, i);
-            if(key > node->keys[i]){
+            if (key > node->keys[i]) {
                 i++;
             }
         }
@@ -91,45 +99,37 @@ void bTree::insertNonFull(bTreeNode* node, int key)
 }
 
 
+// Divide um filho cheio em dois e sobe a chave mediana para o pai
 void bTree::splitChild(bTreeNode* parent, int index)
 {
     bTreeNode* child = parent->children[index];
     bTreeNode* newChild = new bTreeNode(t, child->isLeaf);
 
-    // índice da chave do meio no child
     int mid = child->MiddleKey();
-
-    // guardamos tamanhos originais porque vamos manipular os vetores
     int origKeys = (int)child->keys.size();
     int origChildren = (int)child->children.size();
 
-    // valor que sobe para o parent
     int middleKey = child->keys[mid];
 
-    // 1) copiar as chaves da metade direita para newChild
+    // Copia chaves à direita da mediana para o novo filho
     for (int i = mid + 1; i < origKeys; ++i) {
         newChild->keys.push_back(child->keys[i]);
     }
 
-    // 2) se não for folha, copiar os pointers filhos correspondentes
+    // Se não for folha, copia também os filhos correspondentes
     if (!child->isLeaf) {
-        // os filhos a copiar começam em mid+1 e vão até origChildren-1
         for (int i = mid + 1; i < origChildren; ++i) {
             newChild->children.push_back(child->children[i]);
         }
     }
 
-    // 3) reduzir o tamanho do child esquerdo: ele deve ficar com as primeiras `mid` chaves
-    //    e (se não for folha) com os primeiros (mid+1) filhos
+    // Reduz o tamanho do filho original (esquerdo)
     child->keys.resize(mid);
     if (!child->isLeaf) {
-        child->children.resize(mid + 1); // left child keeps mid+1 children
+        child->children.resize(mid + 1);
     }
 
-    // 4) inserir middleKey no parent na posição correta (index)
     parent->keys.insert(parent->keys.begin() + index, middleKey);
-
-    // 5) inserir newChild como filho imediatamente à direita do child (index + 1)
     parent->children.insert(parent->children.begin() + index + 1, newChild);
 }
 
@@ -140,35 +140,36 @@ void bTree::traverse()
 
 void bTree::traverseNode(bTreeNode* node)
 {
-    if(node != nullptr){
-        int numKeys = node->keys.size();
-        for(int i = 0; i < numKeys; i++){
-            if(!node->isLeaf){
+    if (node != nullptr) {
+        int numKeys = (int)node->keys.size();
+        for (int i = 0; i < numKeys; i++) {
+            if (!node->isLeaf) {
                 traverseNode(node->children[i]);
             }
             std::cout << node->keys[i] << " ";
         }
 
-        if(!node->isLeaf){
+        if (!node->isLeaf) {
             traverseNode(node->children[numKeys]);
         }
     }
 }
 
+// Busca recursiva: retorna o nó que contém a chave, ou nullptr
 bTreeNode* bTree::searchNode(bTreeNode* node, int key)
 {
-    if(node == nullptr){
+    if (node == nullptr) {
         return nullptr;
     }
 
     int i = 0;
-    while (i < node->keys.size() && key > node->keys[i])
+    while (i < (int)node->keys.size() && key > node->keys[i])
         i++;
 
-    if(i < node->keys.size() && node->keys[i] == key)
+    if (i < (int)node->keys.size() && node->keys[i] == key)
         return node;
 
-    if(node->isLeaf)
+    if (node->isLeaf)
         return nullptr;
 
     return searchNode(node->children[i], key);
@@ -194,22 +195,22 @@ void bTree::printTreeNode(bTreeNode* node, int level)
     std::cout << "]\n";
 
     // se não for folha → imprime filhos
-    if (!node->isLeaf)
-    {
+    if (!node->isLeaf) {
         for (bTreeNode* child : node->children)
             printTreeNode(child, level + 1);
     }
 }
 
 
-int main ()
+int main()
 {
+    // Exemplo simples: grau mínimo 3, insere 20 valores aleatórios e imprime
     bTree tree(3);  // grau mínimo t = 3
 
     std::mt19937 rng(std::random_device{}());
     std::uniform_int_distribution<int> dist(-20, 20);
 
-    for(int i = 0; i < 20; i++){
+    for (int i = 0; i < 20; i++) {
         int num = dist(rng);
         std::cout << "Inserting: " << num << std::endl;
         tree.insert(num);
@@ -219,6 +220,6 @@ int main ()
     std::cout << "Traversal of the constructed tree is: ";
     tree.traverse();
     std::cout << std::endl;
-    
+
     return 0;
 }
